+++ src/main/java/cloudwalk/slave/crawler/CrawlController.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package cloudwalk.crawler;
+package cloudwalk.slave.crawler;
 
 
 import redis.clients.jedis.Jedis;
