+++ src/main/java/cloudwalk/slave/crawler/Dupefilter.java
@@ -1,4 +1,4 @@
-package cloudwalk.crawler;
+package cloudwalk.slave.crawler;
 
 import edu.uci.ics.crawler4j.url.URLCanonicalizer;
 import edu.uci.ics.crawler4j.url.WebURL;
