+++ src/test/java/cloudwalk/salve/test/CrawlerTest.java
@@ -1,6 +1,7 @@
+package cloudwalk.salve.test;
 
 
-import cloudwalk.crawler.CrawlController;
+import cloudwalk.slave.crawler.CrawlController;
 import edu.uci.ics.crawler4j.crawler.CrawlConfig;
 import edu.uci.ics.crawler4j.fetcher.PageFetcher;
 import edu.uci.ics.crawler4j.robotstxt.RobotstxtConfig;
